refactor(ilvs): optimize physical graph construction

+++ src/operation/iLVS/source/data_manager/basic/DisjointSet.hpp
@@ -51,7 +51,7 @@ class DisjointSet
 
  private:
   std::vector<int32_t> _parent;
-  std::vector<int32_t> _rank;
+  std::vector<uint8_t> _rank;
 };
 
 }  // namespace ilvs
